% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plankton_data.R
\name{get_plankton_data}
\alias{get_plankton_data}
\title{Gets data from the plankton database}
\usage{
get_plankton_data(sql)
}
\arguments{
\item{sql}{SQL string}
}
\value{
Data frame
}
\description{
Given a query, gets data from Oracle database.
}
\details{
This function will try to use the variable con as a connection to the database.
If con does not exist, it will try to create a new connection using the function open_connection.
}
