% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_samples.R
\name{join_samples}
\alias{join_samples}
\title{Join t_samples with STATIONID}
\usage{
join_samples(qry)
}
\arguments{
\item{qry}{A lazy table with a STATIONID column}
}
\description{
Join t_samples with STATIONID
}
