% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_taxon_groups.R
\name{join_taxon_groups}
\alias{join_taxon_groups}
\title{Join taxon groups from Nivadatabase with a Phytoplankton query}
\usage{
join_taxon_groups(qry)
}
\arguments{
\item{qryPlankton}{A lazy table with a RUBIN_CODE column}
}
\description{
Join taxon groups from Nivadatabase with a Phytoplankton query
}
