% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_connection.R
\name{open_connection}
\alias{open_connection}
\title{Opens a new connection to the Oracle database}
\usage{
open_connection()
}
\value{
A DBI connection object
}
\description{
Opens a new connection to the Oracle database
}
\details{
This function will ask for username, password from the user.
It will look for an installed driver, but otherwise try to get the
driver from the environment variable ORACLE_DRIVER_PATH.
If the environment variable is not set, it will try to use a default path
according to Jupyterhub usage.
}
