% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_plankton_connection.R
\name{open_plankton_connection}
\alias{open_plankton_connection}
\title{Opens a new connection to the Oracle database setting up PHYTOPLANKTON schema}
\usage{
open_plankton_connection()
}
\value{
A DBI connection object
}
\description{
Opens a new connection to the Oracle database setting up PHYTOPLANKTON schema
}
\details{
This function will ask for username, password from the user.
It will look for an installed driver using NIVA common install routines.
It will connect to Nivadatabase production host, but this could be changed by setting
environment variable ORACLE_DNS.
}
