% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_samples.R
\name{pivot_samples}
\alias{pivot_samples}
\title{Pivot Samples to Wide Format
This function pivots a data frame of samples and phytoplankton to wide format.
Each unique sample of date and depth range becomes a separate column.
The BIO_VOLUME values are summed if there are multiple entries for the same sample.
The resulting data frame has columns for STATIONID, RUBIN_CODE, TAXON, and the pivoted sample columns.}
\usage{
pivot_samples(data)
}
\arguments{
\item{data}{A data frame with samples, and optionally phytoplanktons.}
}
\description{
Pivot Samples to Wide Format
This function pivots a data frame of samples and phytoplankton to wide format.
Each unique sample of date and depth range becomes a separate column.
The BIO_VOLUME values are summed if there are multiple entries for the same sample.
The resulting data frame has columns for STATIONID, RUBIN_CODE, TAXON, and the pivoted sample columns.
}
