% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_lakes.R
\name{query_lakes}
\alias{query_lakes}
\title{Queries the t_lakes table}
\usage{
query_lakes(...)
}
\arguments{
\item{...}{Optional filter expressions to apply to the query}
}
\value{
A lazy table of the T_LAKES table
}
\description{
Queries the t_lakes table
}
