% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_taxon_groups.R
\name{query_taxon_groups}
\alias{query_taxon_groups}
\title{Query the taxon groups from the plankton database}
\usage{
query_taxon_groups(...)
}
\arguments{
\item{...}{Optional filter expressions to apply to the query}
}
\value{
A lazy table of the T_TAXON_GROUPS table with columns RUBIN_CODE, GROUP_NAME and GROUP_SORT
}
\description{
Query the taxon groups from the plankton database
}
