% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_taxons.R
\name{query_taxons}
\alias{query_taxons}
\title{Query taxon information from the plankton database}
\usage{
query_taxons()
}
\value{
A lazy table of the T_TAXON_INFORMATION table
}
\description{
Query taxon information from the plankton database
}
