% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_samples.R
\name{query_samples}
\alias{query_samples}
\title{Query samples from the plankton database}
\usage{
query_samples()
}
\value{
A lazy table of the T_SAMPLE table
}
\description{
Query samples from the plankton database
}
